# Optimise HAL_GetTick API