--- a/targets/TARGET_STM/us_ticker.c
+++ b/targets/TARGET_STM/us_ticker.c
@@ -42,9 +42,6 @@ void us_ticker_irq_handler(void);
 // ************************************ 16-bit timer ************************************
 #if TIM_MST_BIT_WIDTH == 16
 
-extern uint32_t prev_time;
-extern uint32_t elapsed_time;
-
 #if defined(TARGET_STM32F0)
 void timer_update_irq_handler(void)
 {
@@ -128,10 +125,6 @@ void init_16bit_timer(void)
 #endif
 
     __HAL_TIM_DISABLE_IT(&TimMasterHandle, TIM_IT_CC1);
-
-    // Used by HAL_GetTick()
-    prev_time = 0;
-    elapsed_time = 0;
 }
 
 // ************************************ 32-bit timer ************************************