<!--
For more information on the requirements for pull requests, please see [the contributing guidelines](https://os.mbed.com/docs/mbed-os/latest/contributing/workflow.html).

NOTE: Do not remove any of the template headings (even for optional sections) as this
template is automatically parsed. 
-->

### Summary of changes <!-- Required -->
_HAL_GetTick_ API has taken about 51us of one-time execution on bare metal profile with small c library (Microlib) on NUCLEO_F303RE target with ARMC6 toolchain and causes failures in lp ticker speed test case from common tickers, so [optimized](https://github.com/ARMmbed/mbed-os/pull/12880#discussion_r417792828)  that API to improve overall performance to achieve optimal execution time and this also improves in other profiles. Refer [PR#12880](https://github.com/ARMmbed/mbed-os/pull/12880) for the more details on this issue.
<!-- 
    Please provide the following information: 

    Description of the the change (what is this fixing / adding / removing?).

    Why the change is needed (if this is fixing a reported issue please summarize what
    the issue is and add the reference. E.g. Fixes #17119).

    NOTE: This section is automatically written to release notes for Feature and 
    Major releases and should contain enough details for a user.
    
-->

#### Impact of changes <!-- Optional -->
with these changes, HAL_GetTick API performance has improved in all profiles, and the below table shows improvement results
| Ticker API  | Full profile(RTOS) <br> (_HAL_GetTick_ **without** optimization) | Bare metal with std lib <br> (_HAL_GetTick_ **without** optimization)| Bare metal profile with small C lib (_HAL_GetTick_ **without** optimization) |All profiles <br> (_HAL_GetTick_ **with** optimization)|
| ------------- | ------------- | ------------ | ------------- | --------------------------|
| clear_interrupt  | 10us| 10us| 51us  | 5us |
| disable_interrupt  |  10us| 10us|51us  | 5us| 
<!-- 
    If there are any implications for users taking this change then they must be 
    provided here. For Major PR types this field is MANDATORY.

    NOTE: This section is automatically written to release notes for Feature and 
    Major releases and should contain enough details for a user.
-->

#### Migration actions required <!-- Optional -->
None.
<!-- 
    This should only be applicable in Major PR types for which this field is MANDATORY.

    NOTE: This section is automatically written to release notes for Feature and 
    Major releases and should contain enough details for a user.
-->

### Documentation <!-- Required -->
None.
<!-- 
    Please provide details of any document updates required, including links to any
    related PRs against the docs repository.
    If no document updates are required please specify 'None', this at least tells us
    that this has been considered.
-->

----------------------------------------------------------------------------------------------------------------
### Pull request type <!-- Required -->

<!--
    Please add only one X to one of the following types. Do not fill multiple types (split the pull request otherwise).
    Please note this is not a GitHub task list, indenting the boxes or changing the format to add a '.' or '*' in front of them would change the meaning incorrectly. 
-->
    [x] Patch update (Bug fix / Target update / Docs update / Test update / Refactor)
    [] Feature update (New feature / Functionality change / New API)
    [] Major update (Breaking change E.g. Return code change / API behaviour change)

----------------------------------------------------------------------------------------------------------------
### Test results <!-- Required -->

<!--
    Provide all the information required, listing all the testing performed. For new targets please attach full test results for all supported compilers.
-->
    [] No Tests required for this change (E.g docs only update)
    [x] Covered by existing mbed-os tests (Greentea or Unittest)
    [] Tests / results supplied as part of this PR
    
    
----------------------------------------------------------------------------------------------------------------
### Reviewers <!-- Optional -->
@evedon @kjbracey-arm 

<!--
    Request additional reviewers with @username or @team
-->

----------------------------------------------------------------------------------------------------------------