@ciarmcom 2020-05-04
@rajkan01, thank you for your changes.
@evedon @kjbracey-arm @ARMmbed/mbed-os-maintainers please review.