@rajkan01 2020-05-05
> It's worth noting in the description that this is also a small optimisation for other C libraries - you had it going from 10us to 5us for ARM_STD, was it?

@kjbracey-arm Updated the PR description to show that information.