@kjbracey-arm 2020-05-06
I think in this case the direct call to `us_ticker_read` isn't the most important factor - using `ticker_read_us` it is actually quite fast on ST because us_ticker is 1MHz, and the ticker_read_us already special-cases that to avoid doing a divisions.

It probably amounts to a few microseconds of the drop (as seen in the 10us->5us change). Still worthwhile. And we bothered to do it in #10609, as it really matters on platforms where it would have needed a division, and we end up avoiding any runtime division (eg wait for an 8MHz timer to increment by 8000, rather than waiting for "timer divided by 8" to increment by 1000).

Here the bulk of the gain is avoiding division, particularly the 64-bit division.

But do check out #12897, which tries to make the general code smaller, and will make it faster in some cases as a secondary benefit. (Smaller when compiled at least - all the ifdefs make the source bigger :( )